# frontend/openapi.json
{"openapi": "3.0.2", "info": {"title": "FastAPI", "version": "0.1.0"}, "servers": [{"url": "http://127.0.0.1:8000", "description": "Local dev server"}], "paths": {"/items/{item_id}": {"get": {"tags": ["items"], "summary": "Read Item", "operationId": "read_item", "parameters": [{"required": true, "schema": {"title": "Item Id", "type": "integer"}, "name": "item_id", "in": "path"}, {"required": false, "schema": {"title": "Q", "type": "string"}, "name": "q", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/code/": {"get": {"tags": ["code"], "summary": "Get File", "operationId": "get_file", "parameters": [{"required": true, "schema": {"title": "File Request Url", "type": "string"}, "name": "file_request_url", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/uploadfile/": {"post": {"tags": ["upload"], "summary": "Create Upload File", "operationId": "create_upload_file", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_upload-create_upload_file"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UploadOutSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/uploadcode/": {"post": {"tags": ["upload"], "summary": "Upload Code", "operationId": "upload_code", "parameters": [{"required": true, "schema": {"title": "Repo Url", "type": "string"}, "name": "repo_url", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RepoUploadOutSchema"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/chat/": {"post": {"tags": ["chat"], "summary": "Send Chat Message", "operationId": "send_chat_message", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatInSchema"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ChatResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"tags": ["root"], "summary": "Main", "operationId": "main", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"Body_upload-create_upload_file": {"title": "Body_upload-create_upload_file", "required": ["file"], "type": "object", "properties": {"file": {"title": "File", "type": "string", "format": "binary"}}}, "ChatInSchema": {"title": "ChatInSchema", "required": ["query", "chat_session", "history"], "type": "object", "properties": {"query": {"title": "Query", "type": "string"}, "chat_session": {"$ref": "#/components/schemas/ChatSessionSchema"}, "history": {"title": "History", "type": "array", "items": {"$ref": "#/components/schemas/ChatMessageSchema"}}}}, "ChatMessageSchema": {"title": "ChatMessageSchema", "required": ["message", "sender"], "type": "object", "properties": {"message": {"title": "Message", "type": "string"}, "sender": {"title": "Sender", "type": "string"}}}, "ChatResponse": {"title": "ChatResponse", "required": ["answer", "source_content", "source_metadata"], "type": "object", "properties": {"answer": {"title": "Answer", "type": "string"}, "source_content": {"title": "Source Content", "type": "string"}, "source_metadata": {"title": "Source Metadata", "type": "object"}}}, "ChatSessionSchema": {"title": "ChatSessionSchema", "required": ["type", "chat_id", "source_id"], "type": "object", "properties": {"type": {"title": "Type", "type": "string"}, "chat_id": {"title": "Chat Id", "type": "string"}, "source_id": {"title": "Source Id", "type": "string"}}}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "RepoInfo": {"title": "RepoInfo", "required": ["source_id", "repo_url", "name"], "type": "object", "properties": {"source_id": {"title": "Source Id", "type": "string"}, "repo_url": {"title": "Repo Url", "type": "string"}, "name": {"title": "Name", "type": "string"}}}, "RepoUploadOutSchema": {"title": "RepoUploadOutSchema", "required": ["type", "repoInfo"], "type": "object", "properties": {"type": {"title": "Type", "type": "string"}, "repoInfo": {"$ref": "#/components/schemas/RepoInfo"}}}, "SourceInfo": {"title": "SourceInfo", "required": ["source_id", "num_bytes", "filename", "display_name"], "type": "object", "properties": {"source_id": {"title": "Source Id", "type": "string"}, "num_bytes": {"title": "Num Bytes", "type": "integer"}, "filename": {"title": "Filename", "type": "string"}, "display_name": {"title": "Display Name", "type": "string"}}}, "UploadOutSchema": {"title": "UploadOutSchema", "required": ["type", "sourceInfo"], "type": "object", "properties": {"type": {"title": "Type", "type": "string"}, "sourceInfo": {"$ref": "#/components/schemas/SourceInfo"}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}